+++ spinnman/connections/abstract_classes/udp_receivers/abstract_udp_eieio_data_receiver.py
@@ -9,8 +9,7 @@ from spinnman.connections.abstract_classes.abstract_eieio_receiver import \
 from spinnman.data.little_endian_byte_array_byte_reader import \
     LittleEndianByteArrayByteReader
 from spinnman.exceptions import SpinnmanTimeoutException, SpinnmanIOException
-from spinnman.messages.eieio.abstract_messages.create_eieio_packets \
-    import create_data_from_reader
+from spinnman.messages.eieio.create_eieio_data import read_eieio_data_message
 
 
 @add_metaclass(ABCMeta)
@@ -59,6 +58,6 @@ class AbstractUDPEIEIODataReceiver(AbstractEIEIOReceiver):
         packet = bytearray(raw_data)
         reader = LittleEndianByteArrayByteReader(packet)
 
-        eieio_packet = create_data_from_reader(reader)
+        eieio_packet = read_eieio_data_message(reader)
 
         return eieio_packet
