+++ spinnman/messages/eieio/data_messages/eieio_16bit_with_payload/eieio_16bit_with_payload_timed_lower_key_prefix_data_message.py
@@ -12,7 +12,7 @@ class EIEIO16BitWithPayloadTimedLowerKeyPrefixDataMessage(
 
         EIEIOWithPayloadDataMessage.__init__(
             self, EIEIOType.KEY_PAYLOAD_16_BIT, prefix=key_prefix,
-            is_time=True, data_reader)
+            is_time=True, data_reader=data_reader)
 
     @staticmethod
     def get_min_packet_length():
